temple of the father exterior terrain change
Added a 'canyon' to railroad players to the bridge area where vital dialogue takes place. Still needs some polish and will probably look a bit different once the lighting is finalized. Feel free to revert if it looks bad. I suggest moving the anomaly cutscene trigger to the temple interior exit so it's not missable.