narrative: remnant test
Added Remnant NPC to trench, along with a subtle distortion effect. Seems to be some transparency conflict with using it as a dialogue NPC, but I can 'fake' the dialogue with an empty actor. Pushing for now so I can solidify tomorrow. Like the keyhole, a nice addition would be some way to simulate depth, or have some additional distortion when looking through the remnants (perhaps some glitchy effect to show this is partly a computer world)