Level2MiddleBlockout(WIP)
This took me about 4 hrs to do, this is just an idea I had for it, if it doesn't work out in y'alls opinion then all good!

This is more or less proof of concept on how to make the level start to work when it comes to making the player look up and also feel like they're in a city.